feat: move tailwind config files to root to sharable between apps
