+++ packages/webb-ui-components/src/css/typography.css
@@ -4,21 +4,21 @@
 
 @font-face {
   font-family: 'BreezeSans';
-  src: url('../fonts/BreezeSans-Regular.ttf') format('truetype');
+  src: url('fonts/BreezeSans-Regular.ttf') format('truetype');
   font-weight: normal;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'BreezeSans';
-  src: url('../fonts/BreezeSans-Medium.ttf') format('truetype');
+  src: url('fonts/BreezeSans-Medium.ttf') format('truetype');
   font-weight: 600;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'BreezeSans';
-  src: url('../fonts/BreezeSans-Bold.ttf') format('truetype');
+  src: url('fonts/BreezeSans-Bold.ttf') format('truetype');
   font-weight: 800;
   font-style: normal;
 }
@@ -27,14 +27,14 @@
 
 @font-face {
   font-family: 'Cousine';
-  src: url('../fonts/Cousine-Regular.ttf') format('truetype');
+  src: url('fonts/Cousine-Regular.ttf') format('truetype');
   font-weight: 400;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'Cousine';
-  src: url('../fonts/Cousine-Bold.ttf') format('truetype');
+  src: url('fonts/Cousine-Bold.ttf') format('truetype');
   font-weight: 800;
   font-style: normal;
 }
